// src/app/Models/chooseBaseClass.ts
// Generated by https://quicktype.io

export interface chooseBaseClass {
    EmpID:           string;
    EmpCode:         string;
    EmpNameC:        string;
    EmpNameE:        string;
    JobID:           number;
    JobName:         string;
    ChiefCode:       string;
    EffectDate:      string;
    DeptaID:         number;
    DeptaName:       string;
    ParentDeptaName: string;
    PosType:         string;
    uiShowBaseString:  string;
}